- reduce clutter in type determination (for you Alon) - moving thing to place
