+++ vips/govips.h
@@ -1,0 +1,6 @@
+#include <stdlib.h>
+#include <vips/vips.h>
+
+#if (VIPS_MAJOR_VERSION < 8)
+  error_requires_version_8
+#endif
