--- vips/type_test.go
@@ -1,30 +1,0 @@
-package vips
-
-import (
-	"io/ioutil"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestJPEG(t *testing.T) {
-	Startup(&Config{})
-
-	buf, err := ioutil.ReadFile(resources + "canyon.jpg")
-	assert.NoError(t, err)
-	assert.NotNil(t, buf)
-
-	imageType := DetermineImageType(buf)
-	assert.Equal(t, ImageTypeJPEG, imageType)
-}
-
-func TestHEIF(t *testing.T) {
-	Startup(&Config{})
-
-	buf, err := ioutil.ReadFile(resources + "citron.heic")
-	assert.NoError(t, err)
-	assert.NotNil(t, buf)
-
-	imageType := DetermineImageType(buf)
-	assert.Equal(t, ImageTypeHEIF, imageType)
-}
